Move existing deploy function into an own deploy folder
