+++ src/contracts/fund/accounting/unit/accounting.test.ts
@@ -1,6 +1,6 @@
 import { Contracts } from '~/Contracts';
 import { initTestEnvironment } from '~/utils/environment/initTestEnvironment';
-import { deployMockSystem } from '~/utils/deployMockSystem';
+import { deployMockSystem } from '~/utils/deploy/deployMockSystem';
 import { BigInteger } from '@melonproject/token-math/bigInteger';
 
 describe('accounting', () => {
