+++ src/contracts/fund/fees/unit/feeManager.test.ts
@@ -1,5 +1,5 @@
 import { initTestEnvironment } from '~/utils/environment/initTestEnvironment';
-import { deployMockSystem } from '~/utils/deployMockSystem';
+import { deployMockSystem } from '~/utils/deploy/deployMockSystem';
 import { deployContract } from '~/utils/solidity/deployContract';
 import { getContract } from '~/utils/solidity/getContract';
 import { Contracts } from '~/Contracts';
