+++ src/contracts/fund/trading/transactions/take0xOrder.test.ts
@@ -4,7 +4,7 @@ import { take0xOrder } from './take0xOrder';
 import { setupInvestedTestFund } from '~/tests/utils/setupInvestedTestFund';
 import { initTestEnvironment } from '~/utils/environment/initTestEnvironment';
 import { getTokenBySymbol } from '~/utils/environment/getTokenBySymbol';
-import { deploySystem } from '~/utils/deploySystem';
+import { deploySystem } from '~/utils/deploy/deploySystem';
 import {
   createOrder,
   signOrder,
